删除firebase账户增加判断

+++ lib/controllers/user_controller.dart
@@ -66,7 +66,9 @@ class UserController extends GetxController {
 
   Future<void> userDelete() async {
     await Request.userDelete();
-    FirebaseAuth.instance.currentUser?.delete();
+    if (GetPlatform.isAndroid) {
+      FirebaseAuth.instance.currentUser?.delete();
+    }
     logout();
   }
 }
